Move HlslKnownTypes to shared generation project
