+++ src/ComputeSharp.D2D1Interop.SourceGenerators/Mappings/HlslKnownMembers.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Numerics;
 using System.Reflection;
 using System.Text.RegularExpressions;
-using ComputeSharp.SourceGenerators.Mappings;
+using ComputeSharp.SourceGeneration.Mappings;
 
 namespace ComputeSharp.D2D1Interop.SourceGenerators.Mappings;
